+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -649,6 +649,13 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     }
 
     /**
+     * Returns the first element having the smallest value according to the provided [comparator] or null if empty.
+     */
+    public fun minWithOrNull(comparator: Comparator<in T>): T? {
+        return if (isEmpty()) null else minWith(comparator)
+    }
+
+    /**
      * @return the first element having the largest value according to the provided [comparator].
      * @throws NoSuchElementException if this ImmutableArray is empty
      */