+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableCharArray.kt
@@ -647,6 +647,13 @@ public value class ImmutableCharArray @PublishedApi internal constructor(
     }
 
     /**
+     * Returns the first element having the smallest value according to the provided [comparator] or null if empty.
+     */
+    public fun minWithOrNull(comparator: Comparator<in Char>): Char? {
+        return if (isEmpty()) null else minWith(comparator)
+    }
+
+    /**
      * @return the first element having the largest value according to the provided [comparator].
      * @throws NoSuchElementException if this ImmutableCharArray is empty
      */