+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableFloatArray.kt
@@ -648,6 +648,13 @@ public value class ImmutableFloatArray @PublishedApi internal constructor(
     }
 
     /**
+     * Returns the first element having the smallest value according to the provided [comparator] or null if empty.
+     */
+    public fun minWithOrNull(comparator: Comparator<in Float>): Float? {
+        return if (isEmpty()) null else minWith(comparator)
+    }
+
+    /**
      * @return the first element having the largest value according to the provided [comparator].
      * @throws NoSuchElementException if this ImmutableFloatArray is empty
      */