+++ immutable-arrays/core/src/test/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableIntArrayTest.kt
@@ -805,6 +805,24 @@ class ImmutableIntArrayTest {
     }
 
     @Test
+    fun `minWithOrNull validation`() {
+        with(emptyImmutableIntArray()) {
+            expectThat(minWithOrNull(Comparator.comparing { it % 3 }))
+                .isNull()
+        }
+
+        with(immutableArrayOf(4)) {
+            expectThat(minWithOrNull(Comparator.comparing { it % 3 }))
+                .isEqualTo(4)
+        }
+
+        with(immutableArrayOf(2, 6, 0)) {
+            expectThat(minWithOrNull(Comparator.comparing { it % 3 }))
+                .isEqualTo(6)
+        }
+    }
+
+    @Test
     fun `maxWith validation`() {
         with(emptyImmutableIntArray()) {
             expectThrows<NoSuchElementException> { maxWith(Comparator.comparing { it % 3 }) }